if final cat or mv fails, fail out

+++ usr.bin/mkdep/mkdep.gcc.sh
@@ -1,6 +1,6 @@
 #!/bin/sh -
 #
-#	$OpenBSD: mkdep.gcc.sh,v 1.6 1997/01/25 14:27:44 niklas Exp $
+#	$OpenBSD: mkdep.gcc.sh,v 1.7 1997/12/12 09:19:59 deraadt Exp $
 #	$NetBSD: mkdep.gcc.sh,v 1.9 1994/12/23 07:34:59 jtc Exp $
 #
 # Copyright (c) 1991, 1993
@@ -77,6 +77,7 @@ if ! mkdir $DTMP ; then
 	echo failed to create tmp dir $DTMP
 	exit 1
 fi
+
 umask $um
 trap 'rm -rf $DTMP ; exit 1' 1 2 3 13 15
 
@@ -94,12 +95,19 @@ fi
 
 if [ $append = 1 ]; then
 	cat $TMP >> $D
-	rm -rf $DTMP
+	if [ $? != 0 ]; then
+		echo 'mkdep: append failed.'
+		rm -rf $DTMP
+		exit 1
+	fi
 else
 	mv $TMP $D
-	rm -rf $DTMP
+	if [ $? != 0 ]; then
+		echo 'mkdep: rename failed.'
+		rm -rf $DTMP
+		exit 1
+	fi
 fi
+
+rm -rf $DTMP
 exit 0
-
-
-
